// src/app/modules/elastic-search/model/common-elastic-search-params.model.ts
import { ElasticSearchDateTimeType } from './elastic-search-date-time-type.model';

export interface CommonElasticSearchParams {
  type?: string;
  size?: number;
  from?: Date;
  to?: Date;
  datetimeType?: ElasticSearchDateTimeType;
  sortFields?: string;
  startIndex?: string;
  includeAggregations?: boolean;
}